GH-124 청산 가격 데이터 읽기 변경

+++ app/src/main/java/com/trueedu/spac/util/ProfitCalculator.kt
@@ -8,7 +8,7 @@ import java.time.temporal.ChronoUnit
  */
 fun redemptionProfitRate(
     currentPrice: Double,
-    redemptionPrice: Int,
+    redemptionPrice: Double,
     targetDate: LocalDate,
 ): Pair<Double?, Double?> {
     val now = LocalDate.now()
